Move list_teams to monitoring_tools
