+++ tests/cli/test_utils.py
@@ -513,35 +513,6 @@ def test_list_instances_no_service(
     assert actual == expected
 
 
-def test_list_teams():
-    fake_team_data = {
-        'team_data': {
-            'red_jaguars': {
-                'pagerduty_api_key': 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-                'pages_slack_channel': 'red_jaguars_pages',
-                'notifications_slack_channel': 'red_jaguars_notifications',
-                'notification_email': 'red_jaguars+alert@yelp.com',
-                'project': 'REDJAGS',
-            },
-            'blue_barracudas': {
-                'pagerduty_api_key': 'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
-                'pages_slack_channel': 'blue_barracudas_pages',
-            },
-        },
-    }
-    expected = {
-        'red_jaguars',
-        'blue_barracudas',
-    }
-    with mock.patch(
-        'paasta_tools.cli.utils._load_sensu_team_data',
-        autospec=True,
-        return_value=fake_team_data,
-    ):
-        actual = utils.list_teams()
-    assert actual == expected
-
-
 def test_lazy_choices_completer():
     completer = utils.lazy_choices_completer(lambda: ['1', '2', '3'])
     assert completer(prefix='') == ['1', '2', '3']
